apps/travel-beat/requirements/development.txt
-r base.txt

# Debug
django-debug-toolbar>=4.2.0
django-extensions>=3.2.3

# Testing
pytest>=8.0.0
pytest-django>=4.7.0
pytest-cov>=4.1.0
factory-boy>=3.3.0

# Code Quality
ruff>=0.2.0
black>=24.1.0
isort>=5.13.2

# Type Hints
mypy>=1.8.0
django-stubs>=4.2.7
